+++ src/containers/viewpaste.tsx
@@ -3,12 +3,13 @@ import { Paste } from "pasty-core";
 import * as React from "react";
 import { connect, Dispatch } from "react-redux";
 
-import config from "../../config";
 import DisplayFile from "../components/displayfile";
+import CopyAndEditIcon from "../components/icons/copyandediticon";
 import CopyShortLinkIcon from "../components/icons/copyshortlinkicon";
 import PasteLoading from "../components/pasteloading";
 
-import { clearPaste, getPaste, setHighlightedLines } from "../actions/creators";
+import config from "../../config";
+import { clearPaste, getPaste, setHighlightedLines, setInitialHighlightedLines } from "../actions/creators";
 import Maybe from "../monads/maybe";
 import { IReducer } from "../reducers/index";
 import { STATE } from "../reducers/paste";
@@ -24,6 +25,7 @@ export interface IViewPasteDispatchProps {
   getPasteAction: (id: string, key: string) => void;
   clearPasteAction: (id: string) => void;
   setHighlightedLines: (index: number, lines: number[]) => void;
+  setInitialHighlightedLines: (index: number, lines: number[]) => void;
 }
 
 export interface IViewPasteProps {
@@ -37,11 +39,17 @@ export interface IViewPasteProps {
       extra?: string,
     },
   };
+  hidden?: boolean;
 }
 
 type PropsType = IViewPasteStateProps & IViewPasteDispatchProps & IViewPasteProps;
 
-class ViewPaste extends React.Component<PropsType, undefined> {
+export class ViewPaste extends React.Component<PropsType> {
+  public static defaultProps: Partial<IViewPasteProps> = {
+    // for loading pastes in the background. Hides the display.
+    hidden: false,
+  };
+
   constructor(props) {
     super(props);
 
@@ -49,15 +57,19 @@ class ViewPaste extends React.Component<PropsType, undefined> {
   }
 
   public componentWillReceiveProps(newProps) {
-    // URL has changed
     if (this.props.match.params.id !== newProps.match.params.id
         || this.props.match.params.key !== newProps.match.params.key) {
         this.updatePaste(newProps);
     }
-  }
 
-  public componentWillUnmount() {
-    this.props.clearPasteAction(this.props.match.params.id);
+    // Once paste loads add highlighted lines.
+    if (newProps.state === STATE.VIEWING && !newProps.paste.isNothing()) {
+      if (newProps.match.params.extra) {
+        newProps.match.params.extra.split(";")
+             .map(this.unserializeLineNumbers)
+             .map((f, i) => newProps.setInitialHighlightedLines(i, f));
+      }
+    }
   }
 
   public render() {
@@ -76,6 +88,10 @@ class ViewPaste extends React.Component<PropsType, undefined> {
       );
     }
 
+    if (this.props.hidden) {
+      return null;
+    }
+
     return (
       <div>
         <h2
@@ -86,6 +102,7 @@ class ViewPaste extends React.Component<PropsType, undefined> {
         <div
           style={{ float: "right" }}
         >
+          <CopyAndEditIcon />
           <CopyShortLinkIcon />
         </div>
 
@@ -99,12 +116,6 @@ class ViewPaste extends React.Component<PropsType, undefined> {
     if (!props.paste.isNothing() && props.paste.getData().name === props.match.params.id
         && props.paste.getData().key === props.match.params.key) {
       return;
-    }
-
-    if (props.match.params.extra) {
-      props.match.params.extra.split(";")
-           .map(this.unserializeLineNumbers)
-           .map((f, i) => props.setHighlightedLines(i, f));
     }
 
     props.getPasteAction(props.match.params.id, props.match.params.key);
@@ -134,16 +145,19 @@ class ViewPaste extends React.Component<PropsType, undefined> {
   }
 }
 
-const mapStateToProps = (state: IReducer, ownProps: IViewPasteProps): IViewPasteStateProps => ({
+export const mapStateToProps = (state: IReducer, ownProps: IViewPasteProps): IViewPasteStateProps => ({
   error: state.messages.general.header !== "",
   paste: state.paste.paste,
   state: state.paste.state,
 });
 
-const mapDispatchToProps = (dispatch: Dispatch<IReducer>): IViewPasteDispatchProps => ({
+export const mapDispatchToProps = (dispatch: Dispatch<IReducer>): IViewPasteDispatchProps => ({
   clearPasteAction: (id: string) => dispatch(clearPaste(id)),
   getPasteAction: (id: string, key: string) => dispatch(getPaste(id, key, `${config.get}${id}`)),
   setHighlightedLines: (index: number, lines: number[]) => dispatch(setHighlightedLines(index, lines)),
+  setInitialHighlightedLines: (index: number, lines: number[]) => (
+    dispatch(setInitialHighlightedLines(index, lines))
+  ),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(ViewPaste);