+++ src/reducers/highlight.ts
@@ -1,5 +1,4 @@
-import range from "lodash/range";
-import uniq from "lodash/uniq";
+import _ = require("underscore");
 
 import {
   ADD_HIGHLIGHTED_LINE,
@@ -24,13 +23,17 @@ const highlight = (state: IHighlightReducer = initial, action) => {
   switch (action.type) {
     case GET_PASTE_FROM_URL:
       return {
-        files: [],
+        files: action.highlight,
       };
 
     case REDIRECT_TO_SUBMITTED_PASTE:
-    case SET_DECRYPTED_PASTE:
       return {
         files: action.paste.files.map(f => []),
+      };
+
+    case SET_DECRYPTED_PASTE:
+      return {
+        files: state.files ? state.files : action.paste.files.map(f => []),
       };
 
     case SET_INITIAL_HIGHLIGHTED_LINES:
@@ -47,10 +50,17 @@ const highlight = (state: IHighlightReducer = initial, action) => {
       const newState = {
         files: [...state.files.map(lines => [...lines])],
       };
-      newState.files[action.index] = [
-        ...state.files[action.index],
-        action.line,
-      ];
+
+      if (!newState.files[action.index].includes(action.line)) {
+        newState.files[action.index] = [
+          ...state.files[action.index],
+          action.line,
+        ];
+      } else {
+        newState.files[action.index].splice(
+          newState.files[action.index].indexOf(action.line), 1,
+        );
+      }
 
       return newState;
     }
@@ -76,10 +86,10 @@ const highlight = (state: IHighlightReducer = initial, action) => {
 export default highlight;
 
 export const rangedSelection = (original: number[], lhs: number, rhs: number): number[] => {
-  const newRange: number[] = (lhs > rhs) ? range(rhs, lhs + 1) : range(lhs, rhs + 1);
+  const newRange: number[] = (lhs > rhs) ? _.range(rhs, lhs + 1) : _.range(lhs, rhs + 1);
   // Reverse the input before calling uniq. Uniq keeps the first occurrance and
   // removes any more. This allows the last clicked (last element) element to remain last.
-  return uniq([
+  return _.uniq([
     ...original,
     ...newRange,
   ].reverse()).reverse();